abstract out renderPlot
